requirements.txt
streamlit>=1.20
pandas
duckdb
matplotlib
openai>=1.0.0
python-dotenv
